Adding leEq, greater, and grEq to the signature of ordered fields
The definition of leEq varies from structure to structure.  In order for the
theorems that use [<=] to apply to the usual definition of leEq for the
structure, it is necessary to make leEq part of the signature of the ordered
fields structure.

The new signature for is_COrdField is:

Record is_COrdField (F : CField)
  (less : CCSetoid_relation F) (leEq : Relation F)
  (greater : CCSetoid_relation F) (grEq : Relation F) : CProp :=
  {ax_less_strorder  : strictorder less;
   ax_plus_resp_less : forall x y, less x y -> forall z, less (x[+]z) (y[+]z);
   ax_mult_resp_pos  : forall x y, less Zero x -> less Zero y -> less Zero (x[*]y);
   ax_less_conf_ap   : forall x y, Iff (x [#] y) (less x y or less y x);
   def_leEq : forall x y, (leEq x y) <-> (Not (less y x));
   def_greater : forall x y, Iff (greater x y) (less y x);
   def_grEq : forall x y, (grEq x y) <-> (leEq y x)}.

I have added leEq_def, greater_def, and grEq_def.  These are unfolded version of
the definitions in the structure.

I have also added default_greater, default_leEq, and default_grEq definition
that give the former definitions to assist migration to this new signature.

The lemmas approach_zero and approach_zero_weak have changed to the following
(previously) convertable types---the types were convertible before this patch,
but now the types are no longer convertable.

-Lemma approach_zero : forall x : R, (forall e, Zero [<] e -> x [<] e) -> Not (Zero [<] x).
+Lemma approach_zero : forall x : R, (forall e, Zero [<] e -> x [<] e) -> x [<=] Zero.

-Lemma approach_zero_weak : forall x : R, (forall e, Zero [<] e -> x [<=] e) -> Not (Zero [<] x).
+Lemma approach_zero_weak : forall x : R, (forall e, Zero [<] e -> x [<=] e) -> x [<=] Zero.

These lemmas were always used on goals with leEq.

This change breaks backwards compatibility moderately.  It affects the use of
auto, simpl, etc. and it is possible the users will have to change their proofs
in unpredictable ways. The most common change is to replace unfold leEq with
rewrite leEq_def.

- unfold leEq in H5.
+ rewrite leEq_def in H5.

However several other types of changes may be necessary for users of CoRN.  Here
are a few common diffs.

* rewriting doesn't work under foralls, so sometimes rewriting must be done
later than the unfolding was.

-unfold leEq in |- *.
 intros x y z H H0.
+rewrite leEq_def in *.

* and some times a bit more work needs to be done.

-apply (H N).
+pose (c:=H N).
+rewrite leEq_def in c.
+apply c.

* folding is done by rewrite <- leEq_def.

- change (Zero [<=] x) in |- *.
+ change (Not (x [<] Zero)).
+ rewrite <- leEq_def.

* inserting a rewrite step before intros.

+rewrite leEq_def in *.
 intro.

* inserting a rewrite step before application.

+rewrite leEq_def in H.
 elim (H H1).

* replacing auto with firstorder with leEq_def

-auto.
+firstorder with leEq_def.

* or adding a rewrite before auto

-apply nexp_resp_less; auto.
+apply nexp_resp_less; try rewrite leEq_def; auto.

* Rewriting currently doesn't work for IFF, so it is more difficult to replace
unfold greater.  Fortunately greater is rarely used in CoRN, and I personally
recommend against using it.

-unfold greater in H1.
+pose (H1':=(CAnd_proj1 _ _ (greater_def _ _ _) H1)).
+clearbody H1'.
+clear H1.
+rename H1' into H1.

To this end, I've removed some uses of greater by changing the types of
strict_dec_glues and strict_dec_glues' to the following (previously) convertable
types.

 Lemma strict_dec_glues : forall a b c F (Hab : a [<=] b) (Hbc : b [<=] c) (Hac : a [<=] c),
  included (Compact Hac) (Dom F) ->
- (forall x y, Compact Hab x -> Compact Hab y -> x[>]y -> forall Hx Hy, F x Hx [<] F y Hy) ->
- (forall x y, Compact Hbc x -> Compact Hbc y -> x[>]y -> forall Hx Hy, F x Hx [<] F y Hy) ->
- forall x y, Compact Hac x -> Compact Hac y -> x[>]y -> forall Hx Hy, F x Hx [<] F y Hy.
+ (forall x y, Compact Hab x -> Compact Hab y -> y[<]x -> forall Hx Hy, F x Hx [<] F y Hy) ->
+ (forall x y, Compact Hbc x -> Compact Hbc y -> y[<]x -> forall Hx Hy, F x Hx [<] F y Hy) ->
+ forall x y, Compact Hac x -> Compact Hac y -> y[<]x -> forall Hx Hy, F x Hx [<] F y Hy.

 Lemma strict_dec_glues' : forall a b c F, a [<] b -> b [<] c -> included (olor a c) (Dom F) ->
- (forall x y, olcr a b x -> olcr a b y -> x[>]y -> forall Hx Hy, F x Hx [<] F y Hy) ->
- (forall x y, clor b c x -> clor b c y -> x[>]y -> forall Hx Hy, F x Hx [<] F y Hy) ->
- forall x y, olor a c x -> olor a c y -> x[>]y -> forall Hx Hy, F x Hx [<] F y Hy.
+ (forall x y, olcr a b x -> olcr a b y -> y[<]x -> forall Hx Hy, F x Hx [<] F y Hy) ->
+ (forall x y, clor b c x -> clor b c y -> y[<]x -> forall Hx Hy, F x Hx [<] F y Hy) ->
+ forall x y, olor a c x -> olor a c y -> y[<]x -> forall Hx Hy, F x Hx [<] F y Hy.

git-archimport-id: r.oconnor@cs.ru.nl--Nijmegen/CoRN--roconnor-dev--3--patch-6