Adding lemmas about ArcTangent (plus a few others).
I'm adding a series of lemmas about Arctangent in a file named
tranc/MoreArcTan.v.  In additon to small colletion of lemmas on
Arctangent, I've added a few other lemmas about trigonometirc functions.
Mostly lemmas on how to prove something is in the domain of Tangent, but
also how to show that Sin x[#]Zero and Cos x[#]Zero.

I also added a lemma showing the geometric series converges on (-1,1).
This is then used to show that arctangent's power series converges to
arctangent on (-1,1).


git-archimport-id: r.oconnor@cs.ru.nl--Nijmegen/CoRN--roconnor-dev--3--patch-46