Add sketch of new abstract interface for integration.
This new formulation no longer starts with an integration operator whose behavior is then described when applied to certain functions (this is the Bishop formulation). The basic problem with this approach is that specifying the properties this way means that one must already have some domain in which the integration operator can be used, which gets hairy if one takes possible domain restrictions seriously. The new formulation is much simpler and more to the point, because it only specifies what the result of integrating a single function on a single range must look like (namely Riemann sums). There is no greater context or implied ability to integrate anything else; all that stuff is now easily expressible in terms of this primitive.