Remove support for untested versions of Coq.
I am making this suggestion because e.g. when I tried to compile CoRN
from source with Coq 8.12.x, I had many problems that had to be fixed by
manually adjusting proofs. It looks to me as if the Docker CI that CoRN
has configured is only valid for 8.14-8.17 and dev.

I'm an outsider to this project so it's possible this change doesn't
make sense or ignores some considerations--if so I'd be interested to
learn what these are.