Use `bc` for version comparison in configure.sh
Because versions have the form X.Y.Z (major/minor/patch), the regex used
to grab the minor.patch version for comparison yields a float. The -gt
flag does not process floats, but `bc`, which is broadly available on
*nix systems, does.

Alternatively since support for versions less than 8.14 is dropped, we
could just delete this conditional.