Simple deprecations of lt_n_Sm_le.
The notation lt_n_Sm_le is deprecated for Nat.lt_succ_r, but
Nat.lt_succ_r is not a drop-in replacement for lt_n_Sm_le.
This means that e.g. in "./algebra/Bernstein.v", replacing lt_n_Sm_le is
not so easy, because lt_n_Sm_le is used as an argument to other proof
terms.